:boom: Refactored cookiecutter template.

+++ stack.{{cookiecutter.project_slug}}/scripts/clean.sh
@@ -10,10 +10,19 @@ cd "${_PROJECT_DIR}" || exit 2
 # Loading base script:
 # shellcheck disable=SC1091
 source ./scripts/base.sh
+
+# Loading .env file (if exists):
+if [ -f ".env" ]; then
+	# shellcheck disable=SC1091
+	source .env
+fi
 ## --- Base --- ##
 
 
 ## --- Variables --- ##
+# Load from envrionment variables:
+PROJECT_SLUG="${PROJECT_SLUG:-{{cookiecutter.project_slug}}}"
+
 # Flags:
 _IS_ALL=false
 ## --- Variables --- ##
@@ -39,6 +48,12 @@ main()
 	## --- Menu arguments --- ##
 
 
+	if docker compose ps | grep 'Up' > /dev/null 2>&1; then
+		echoWarn "Docker is running, please stop it before cleaning."
+		exit 1
+	fi
+
+
 	echoInfo "Cleaning..."
 
 	find . -type f -name ".DS_Store" -print -delete || exit 2
@@ -49,8 +64,19 @@ main()
 	find . -type d -name ".pytest_cache" -exec rm -rfv {} + || exit 2
 	# find . -type d -name ".git" -prune -o -type d -name "logs" -exec rm -rfv {} + || exit 2
 
+	rm -rfv "./${PROJECT_SLUG}" || exit 2
+	rm -rfv "./volumes/storage/${PROJECT_SLUG}/logs" || exit 2
+	# rm -rfv ./app/logs || exit 2
+	# rm -rfv ./logs || exit 2
+
 	if [ "${_IS_ALL}" == true ]; then
-		rm -rfv ./stack.fastapi-orm-template || exit 2
+		rm -rf "./volumes/.vscode-server/*" || exit 2
+		rm -rfv ./volumes/storage/postgresql/data || exit 2
+		rm -rfv "./volumes/storage/postgresql/logs/*" || exit 2
+		rm -rfv "./volumes/storage/${PROJECT_SLUG}/data" || exit 2
+		rm -rfv ./volumes/backups || exit 2
+
+		docker compose down -v || exit 2
 	fi
 
 	echoOk "Done."
